finished search for employees name or deparment id, created policies allowing only user Ropioner to CRUD countries

+++ app/Http/Controllers/Backend/CountryController.php
@@ -11,6 +11,7 @@ class CountryController extends Controller
 {
     public function index(Request $request)
     {
+
         $countries = Country::all();
 
         if($request->has('search')){
@@ -23,11 +24,16 @@ class CountryController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Country::class);
+
         return view('country.create');
     }
 
     public function store(CountryStoreRequest $request)
     {
+
+        $this->authorize('create', Country::class);
+
         Country::create($request->validated());
 
         return redirect()->route('country.index')->with('message', 'Country Created Successfully');
@@ -35,17 +41,23 @@ class CountryController extends Controller
 
     public function edit(Country $country)
     {
+        $this->authorize('update', Country::class);
         return view('country.edit', compact('country'));
     }
 
     public function update(CountryStoreRequest $request, Country $country)
     {
+        $this->authorize('update', Country::class);
+        
         $country->update($request->validated());
         return redirect()->route('country.index')->with('message', 'Country Updated Successfully');
     }
 
     public function destroy(Country $country)
     {
+
+        $this->authorize('delete', Country::class);
+
         $country->delete();
 
         return redirect()->route('country.index')->with('message', 'Country Deleted Successfully');
